BioLens/requirements.txt
flask
google-cloud-vision
python-multipart
torch
transformers
pillow
